changes to make it easier to autogen address
