+++ google/data_source_google_compute_address.go
@@ -2,9 +2,16 @@ package google
 
 import (
 	"fmt"
+	"regexp"
 	"strconv"
+	"strings"
 
 	"github.com/hashicorp/terraform/helper/schema"
+)
+
+var (
+	computeAddressIdTemplate = "projects/%s/regions/%s/addresses/%s"
+	computeAddressLinkRegex  = regexp.MustCompile("projects/(.+)/regions/(.+)/addresses/(.+)$")
 )
 
 func dataSourceGoogleComputeAddress() *schema.Resource {
@@ -73,4 +80,64 @@ func dataSourceGoogleComputeAddressRead(d *schema.ResourceData, meta interface{}
 
 	d.SetId(strconv.FormatUint(uint64(address.Id), 10))
 	return nil
+}
+
+type computeAddressId struct {
+	Project string
+	Region  string
+	Name    string
+}
+
+func (s computeAddressId) canonicalId() string {
+	return fmt.Sprintf(computeAddressIdTemplate, s.Project, s.Region, s.Name)
+}
+
+func parseComputeAddressId(id string, config *Config) (*computeAddressId, error) {
+	var parts []string
+	if computeAddressLinkRegex.MatchString(id) {
+		parts = computeAddressLinkRegex.FindStringSubmatch(id)
+
+		return &computeAddressId{
+			Project: parts[1],
+			Region:  parts[2],
+			Name:    parts[3],
+		}, nil
+	} else {
+		parts = strings.Split(id, "/")
+	}
+
+	if len(parts) == 3 {
+		return &computeAddressId{
+			Project: parts[0],
+			Region:  parts[1],
+			Name:    parts[2],
+		}, nil
+	} else if len(parts) == 2 {
+		// Project is optional.
+		if config.Project == "" {
+			return nil, fmt.Errorf("The default project for the provider must be set when using the `{region}/{name}` id format.")
+		}
+
+		return &computeAddressId{
+			Project: config.Project,
+			Region:  parts[0],
+			Name:    parts[1],
+		}, nil
+	} else if len(parts) == 1 {
+		// Project and region is optional
+		if config.Project == "" {
+			return nil, fmt.Errorf("The default project for the provider must be set when using the `{name}` id format.")
+		}
+		if config.Region == "" {
+			return nil, fmt.Errorf("The default region for the provider must be set when using the `{name}` id format.")
+		}
+
+		return &computeAddressId{
+			Project: config.Project,
+			Region:  config.Region,
+			Name:    parts[0],
+		}, nil
+	}
+
+	return nil, fmt.Errorf("Invalid compute address id. Expecting resource link, `{project}/{region}/{name}`, `{region}/{name}` or `{name}` format.")
 }
